(integration) - Add DocuSign integration support.
+ Added functions to mimic JWT validation from docusign-esign library.
+ DocuSign's GraphQL endpoints offer an interface to generate access tokens to call other endpoints.

** Tech debt **
It's important to keep studying and understanding the different DocuSign workflows in production. Currently there's no explicit proof that JWT authentication is the way to go to finish this integration. At least as PoC it will suffice, but a meeting with a DocuSign representative would help us understand better which would be a better approach.